mysite/requirements.txt
Django==4.1.3
djangorestframework==3.14.0
Pillow==9.0.1
pipenv==11.9.0
requests==2.25.1
virtualenv==20.16.6
